+++ addons/web/models/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import ir_qweb
+from . import ir_qweb_fields
 from . import ir_http
 from . import ir_ui_menu
 from . import models
