+++ addons/web_editor/models/__init__.py
@@ -2,7 +2,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import ir_attachment
-from . import ir_qweb
+from . import ir_qweb_fields
 from . import ir_ui_view
 from . import ir_http
 from . import ir_translation
