+++ addons/web_editor/tests/test_converter.py
@@ -8,7 +8,7 @@ from lxml.builder import E
 
 from odoo.tests import common
 from odoo.tests.common import BaseCase
-from odoo.addons.web_editor.models.ir_qweb import html_to_text
+from odoo.addons.web_editor.models.ir_qweb_fields import html_to_text
 
 
 class TestHTMLToText(BaseCase):
