+++ addons/web_unsplash/models/__init__.py
@@ -3,4 +3,4 @@
 
 from . import res_config_settings
 from . import res_users
-from . import ir_qweb
+from . import ir_qweb_fields
